backend/requirements.txt
flask
werkzeug
flask_cors
sqlalchemy
requests
python-dotenv
beautifulsoup4
pytest
requests-mock
